remove files in input/ src/ mfront-law/ and mesh/ + add options for markdown_reader
