+++ ex6/CMakeLists.txt
@@ -37,12 +37,13 @@ if (NOT TARGET mfem-mgis::MFEMMGIS)
   find_package(MFEMMGIS REQUIRED)
 endif(NOT TARGET mfem-mgis::MFEMMGIS)
 
+SET(HEARDERS include)
 # Require timers
 if(MFEM_USE_MPI)
 	add_compile_options(-D__MPI)
-	set(TIMERS timer.cpp timer.hpp)
+	set(TIMERS src/common/timer.cpp include/common/timer.hpp)
 	add_library(timers SHARED ${TIMERS})
-	target_include_directories(timers PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+	target_include_directories(timers PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
 	find_package(MPI REQUIRED)
 	target_link_libraries(timers PUBLIC MPI::MPI_CXX)
 endif()
@@ -51,15 +52,16 @@ endif()
 
 if(MFEM_USE_MPI)
 
-	SET(utiles test_parameters.cpp solver_name.cxx precond_name.cxx data_gathering.cxx common.cxx)
-	add_library(configuration STATIC ${utiles})
-	target_link_libraries(configuration mfem-mgis::MFEMMGIS)
-	target_link_libraries(configuration timers)
-	target_include_directories(configuration
+	SET(HEARDERS "include")
+	SET(utiles src/parameters/test_parameters.cpp src/solver/solver_name.cxx src/solver/precond_name.cxx src/common/data_gathering.cxx src/common/common.cxx src/parameters/catch_functions.cxx)
+	add_library(labo-tools STATIC ${utiles})
+	target_link_libraries(labo-tools mfem-mgis::MFEMMGIS)
+	target_link_libraries(labo-tools timers)
+	target_include_directories(labo-tools
 		SYSTEM  
 		PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
 		PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
-		PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}"
+		PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/include"
 		)
 
 
@@ -68,13 +70,13 @@ if(MFEM_USE_MPI)
 	endfunction()
 
 	function(build_case name)
-		add_library(${name} STATIC ${name}.cxx)
+		add_library(${name} STATIC src/test-cases/${name}.cxx)
 		target_link_libraries(${name} mfem-mgis::MFEMMGIS)
 		target_include_directories(${name}
 			SYSTEM  
 			PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
 			PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
-			PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}"
+			PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/include"
 			)
 	endfunction()
 
@@ -84,43 +86,41 @@ if(MFEM_USE_MPI)
 
 	# Define the executable to build for this example
 	add_executable(perf perf.cxx)
-	target_link_libraries(perf configuration)
+	target_link_libraries(perf labo-tools)
 	add_case(perf cas_cible_1)
 	add_case(perf fissuration)
 
 	add_executable(reader read_markdown.cxx)
-	target_link_libraries(reader configuration)
+	target_link_libraries(reader labo-tools)
 
 	add_executable(one_test one_test.cxx)
-	target_link_libraries(one_test configuration)
+	target_link_libraries(one_test labo-tools)
 
 	add_case(one_test cas_cible_1)
 	add_case(one_test fissuration)
 
 	add_executable(petsc_file_generation use_petsc.cxx)
-	target_link_libraries(petsc_file_generation configuration)
+	target_link_libraries(petsc_file_generation labo-tools)
 
 
 	# Copy needed file for execution
 	file(COPY
-		${PROJECT_SOURCE_DIR}/Elasticity.mfront
-		${PROJECT_SOURCE_DIR}/AT1MicromorphicDamage.mfront
-		${PROJECT_SOURCE_DIR}/MicromorphicDamageI.mfront
-		${PROJECT_SOURCE_DIR}/cube_2mat_per.mesh
+		${PROJECT_SOURCE_DIR}/mfront-law
+		${PROJECT_SOURCE_DIR}/mesh
 		DESTINATION ${PROJECT_BINARY_DIR})
 
 
 	# Use mfront to generate libBehaviour.so
 	add_custom_command(
 		OUTPUT "src/libBehaviour-elasticity.so"
-		COMMAND mfront --obuild --interface=generic ${PROJECT_BINARY_DIR}/Elasticity.mfront
-		DEPENDS ${PROJECT_BINARY_DIR}/Elasticity.mfront
+		COMMAND mfront --obuild --interface=generic ${PROJECT_BINARY_DIR}/mfront-law/Elasticity.mfront
+		DEPENDS ${PROJECT_BINARY_DIR}/mfront-law/Elasticity.mfront
 				COMMENT "Performing mfront code generation for cas cible 1")
 	
 	add_custom_command(
 		OUTPUT "src/libBehaviour-for-fissuration.so"
-		COMMAND mfront --obuild --interface=generic ${PROJECT_BINARY_DIR}/AT1MicromorphicDamage.mfront ${PROJECT_BINARY_DIR}/MicromorphicDamageI.mfront
-		DEPENDS ${PROJECT_BINARY_DIR}/AT1MicromorphicDamage.mfront ${PROJECT_BINARY_DIR}/MicromorphicDamageI.mfront
+		COMMAND mfront --obuild --interface=generic ${PROJECT_BINARY_DIR}/mfront-law/AT1MicromorphicDamage.mfront ${PROJECT_BINARY_DIR}/mfront-law/MicromorphicDamageI.mfront
+		DEPENDS ${PROJECT_BINARY_DIR}/mfront-law/AT1MicromorphicDamage.mfront ${PROJECT_BINARY_DIR}/mfront-law/MicromorphicDamageI.mfront
 		COMMENT "Performing mfront code generation for fissuration")
 
 		add_custom_target(generate-libBehaviour-cas-cible-1 ALL
@@ -134,18 +134,18 @@ if(MFEM_USE_MPI)
 	# Define a test that will be called during the "make check"
 	function(add_perf)
 		set(test "perf")
-		add_test(NAME ${test}_cube_2mat_per COMMAND mpirun -n 3 ./perf --mesh cube_2mat_per.mesh) #see later why it fails with -n 4 and it works with n = 1,2,3,5,6,7,44
+		add_test(NAME ${test}_cube_2mat_per COMMAND mpirun -n 3 ./perf --mesh mesh/cube_2mat_per.mesh) #see later why it fails with -n 4 and it works with n = 1,2,3,5,6,7,44
 	endfunction(add_perf)
 
 	function(add_solver_preconditioner)
 		#no foreach
-		add_test(NAME periodic_hypreCG_HypreDiagScale_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 0 --mesh cube_2mat_per.mesh) 
-		add_test(NAME periodic_hypreCG_HypreBoomerAMG_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 3 --mesh cube_2mat_per.mesh) 
-		add_test(NAME periodic_hypreCG_HypreILU_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 4 --mesh cube_2mat_per.mesh) 
-		add_test(NAME periodic_hypreCG_any_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 5 --mesh cube_2mat_per.mesh) 
-		add_test(NAME periodic_CGSolver_HypreDiagScale_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 0 --mesh cube_2mat_per.mesh) 
-		add_test(NAME periodic_CGSolver_HypreParaSails_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 1 --mesh cube_2mat_per.mesh) 
-		add_test(NAME periodic_CGSolver_any_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 5 --mesh cube_2mat_per.mesh) 
+		add_test(NAME periodic_hypreCG_HypreDiagScale_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 0 --mesh mesh/cube_2mat_per.mesh) 
+		add_test(NAME periodic_hypreCG_HypreBoomerAMG_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 3 --mesh mesh/cube_2mat_per.mesh) 
+		add_test(NAME periodic_hypreCG_HypreILU_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 4 --mesh mesh/cube_2mat_per.mesh) 
+		add_test(NAME periodic_hypreCG_any_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 5 --mesh mesh/cube_2mat_per.mesh) 
+		add_test(NAME periodic_CGSolver_HypreDiagScale_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 0 --mesh mesh/cube_2mat_per.mesh) 
+		add_test(NAME periodic_CGSolver_HypreParaSails_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 1 --mesh mesh/cube_2mat_per.mesh) 
+		add_test(NAME periodic_CGSolver_any_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 5 --mesh mesh/cube_2mat_per.mesh) 
 	endfunction(add_solver_preconditioner)
 
 	add_perf()
