+++ ex6/include/parameters/catch_functions.hxx
@@ -1,0 +1,28 @@
+#pragma once
+
+#include <solver/solver_name.hxx>
+#include <solver/precond_name.hxx>
+
+// get test case (kernel, solver, preconditionner and match function)
+#include <test-cases/cas_cible_1.hxx>
+#include <test-cases/fissuration.hxx>
+
+
+
+template<typename Solver, typename Pc>
+std::function<void(const TestParameters&, const bool, const Solver, const Pc, gather_information&)> get_kernel(int a_case)
+{
+	switch(a_case)
+	{
+		case 1: return cas_cible_1::kernel<Solver,Pc>;
+		case 2: return fissuration::kernel<Solver,Pc>;
+		default: return cas_cible_1::kernel<Solver,Pc>;
+	}
+}
+
+std::vector<solver_name> get_solvers(int a_case);
+std::vector<petsc_ksp_type> get_solvers_with_petsc(int a_case);
+std::vector<precond_name> get_pc(int a_case);
+std::vector<petsc_pc_type> get_pc_with_petsc(int a_case);
+std::function<bool(solver_name,precond_name)> get_match(int a_case);
+std::function<bool(petsc_ksp_type,petsc_pc_type)> get_match_with_petsc(int a_case);
