+++ ex6/include/solver/solver_name.hxx
