+++ ex6/mesh/cube_2mat_per.mesh
