+++ ex6/mfront-law/AT1MicromorphicDamage.mfront
