+++ ex6/mfront-law/Elasticity.mfront
