+++ ex6/mfront-law/MicromorphicDamageI.mfront
