+++ ex6/src/common/common.cxx
@@ -1,4 +1,4 @@
-#include <common.hxx>
+#include <common/common.hxx>
 namespace common 
 {
 	void print_statistics(std::string string_solver, std::string string_pc, double time)
