+++ ex6/src/common/timer.cpp
@@ -1,4 +1,4 @@
-#include<timer.hpp>
+#include<common/timer.hpp>
 #ifdef __MPI
 #include"mpi.h"
 #endif
